pyproject.toml
[tool.poetry]
name = "flask-line-bot-heroku"
version = "0.1.0"
description = ""
authors = ["hsuanchi <lawrencewang0321@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Flask = "^1.1.2"
line-bot-sdk = "^1.17.0"
gunicorn = "^20.0.4"
PyMySQL = "^1.0.2" 
python-geoip = "^1.2"
[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
